requirements.txt
# python3.11
streamlit==1.28.2
openai
streamlit-lottie
streamlit_folium
python-dotenv
pillow